+++ tests/LaravelModuleTest.php
@@ -154,8 +154,8 @@ class ModuleTest extends BaseTestCase
     /** @test */
     public function it_checks_module_enabled_status()
     {
-        $this->assertFalse($this->module->enabled());
-        $this->assertTrue($this->module->disabled());
+        $this->assertFalse($this->module->isEnabled());
+        $this->assertTrue($this->module->isDisabled());
     }
 
     /** @test */